+++ test/IO/test_io.py
@@ -119,6 +119,7 @@ binary_example_file_list = _convert_filelist(['di-1.di',
                                               'poir-1.poir',
                                               'stp-1.stp',
                                               'top-1.top',
+                                              'plux-1.plux',
                                               # MPI I/O does not support Python streams
                                               ] + ([] if NuMPI._has_mpi4py else ['example-2d.npy']))
 
@@ -654,6 +655,7 @@ def test_detect_format(file_format_examples):
     assert detect_format(os.path.join(file_format_examples, 'poir-1.poir')) == 'poir'
     assert detect_format(os.path.join(file_format_examples, 'stp-1.stp')) == 'wsxm'
     assert detect_format(os.path.join(file_format_examples, 'top-1.top')) == 'wsxm'
+    assert detect_format(os.path.join(file_format_examples, 'plux-1.plux')) == 'plux'
 
 
 def test_to_matrix():