+++ SurfaceTopography/ChangeLog.md
@@ -5,6 +5,7 @@ v1.10.0 (not yet released)
 --------------------------
 
 - ENH: Reader for WSxM (.stp, .top) files
+- ENH: Reador for Sensorfar XML SPM (.plux) files
 
 v1.9.0 (28Aug23)
 ----------------